+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,10 +1,13 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 452
+INVENTREE_API_VERSION = 453
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
+
+v453 -> 2026-02-12 : https://github.com/inventree/InvenTree/pull/10983
+    - Adds an API to gather dynamic price information for a Part
 
 v452 -> 2026-02-10 : https://github.com/inventree/InvenTree/pull/11276
     - Adds "install_into_detail" field to the BuildItem API endpoint
@@ -24,6 +27,7 @@ v448 -> 2026-02-05 : https://github.com/inventree/InvenTree/pull/11257
 
 v447 -> 2026-02-02 : https://github.com/inventree/InvenTree/pull/11242
     - Adds "sub_part_active" filter to BomItem API endpoint
+>>>>>>> master
 
 v446 -> 2026-02-01 : https://github.com/inventree/InvenTree/pull/11232
     - Allow ordering of test results by started_datetime and finished_datetime fields
@@ -82,6 +86,7 @@ v431 -> 2025-12-14 : https://github.com/inventree/InvenTree/pull/11006
     - Remove duplicate "address" field on the Company API endpoint
     - Make "primary_address" field optional on the Company API endpoint
     - Remove "address_count" field from the Company API endpoint
+>>>>>>> 612e54b415d09686a450f6025b6787da4d91552e
 
 v430 -> 2025-12-04 : https://github.com/inventree/InvenTree/pull/10699
     - Removed the "PartParameter" and "PartParameterTemplate" API endpoints